Move the tests for /settings to /session/settings
